requirements.txt
pytesseract
pillow
requests
bs4
